+++ src/Http/Requests/UpdateShippingMethod.php
@@ -24,6 +24,7 @@ class UpdateShippingMethod extends FormRequest implements UpdateShippingMethodCo
         return [
             'name' => 'required|min:2|max:255',
             'carrier_id' => 'required|exists:carriers,id',
+            'zone_id' => 'sometimes|nullable|exists:zones,id',
             'configuration' => 'sometimes|json',
             'is_active' => 'sometimes|boolean',
         ];